+++ app/Http/Controllers/Budget/EditController.php
@@ -103,7 +103,7 @@ class EditController extends Controller
         ];
         if ($autoBudget) {
             $amount                          = $hasOldInput ? $request->old('auto_budget_amount') : $autoBudget->amount;
-            $preFilled['auto_budget_amount'] = number_format((float)$amount, $autoBudget->transactionCurrency->decimal_places);
+            $preFilled['auto_budget_amount'] = number_format((float)$amount, $autoBudget->transactionCurrency->decimal_places, '.', '');
         }
 
         // put previous url in session if not redirect from store (not "return_to_edit").