+++ config/firefly.php
@@ -101,8 +101,8 @@ return [
         'webhooks'     => true,
         'handle_debts' => true,
     ],
-    'version'                      => '5.6.5',
-    'api_version'                  => '1.5.4',
+    'version'                      => '5.6.6',
+    'api_version'                  => '1.5.5',
     'db_version'                   => 18,
 
     // generic settings